pyproject.toml
[project]
name = "rag-genai-dbs-codex"
version = "0.1.0"
description = "RAG application with GenAI and vector databases using ColPali and Qdrant"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.31.1",
    "qdrant-client>=1.15.1",
    "llama-index-core>=0.10.12",
    "llama-index-embeddings-ollama>=0.1.2",
    "torch>=2.5.0,<2.8.0",
    "torchvision>=0.20.0",
    "colpali-engine==0.3.12",
    "transformers>=4.53.1,<4.54.0",
    "accelerate>=0.26.1",
    "qwen-vl-utils>=0.0.8",
    "pdf2image>=1.17.0",
    "Pillow>=10.2.0",
    "requests>=2.31.0",
    "tqdm>=4.66.1",
    "numpy>=1.26.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[dependency-groups]
dev = []

[tool.uv.sources]
